Cast step frame interval to integer
This was needed to avoid a type error in QTimer.start() in my new setup:
Python 3.10.12, PyQt 5.15.6.

+++ canvas.py
@@ -572,7 +572,7 @@ class DomainCanvas(app.Canvas):
     # initialize step animation
     self.step = lambda t: identity(3)
     self.step_frame_rate = 30
-    self.step_frame_interval = 1000 / self.step_frame_rate
+    self.step_frame_interval = int(1000 / self.step_frame_rate)
     self.step_speed = 6
     self.step_nframes = 1
     self.step_timer = QTimer()
